Fun2Learn-AR_app
it is app for kids teaching. It is built with Unity 3D as a game engine and targeting Android platform. Using the technology of Augmented reality. It makes learning much easier and funnier. This is the first phase which is concerned with alphabets and numbers in both languages Arabic and English.